pyproject.toml
[tool.poetry]
name = "sitn-django-plugins"
version = "1.0.0"
description = "Development repository for custom sitn plugins"
authors = ["maltaesousa <stephane.maltaesousa@ne.ch>"]
license = "BSD"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2"
psycopg2 = "^2.9.10"
twine = "^6.1.0"
build = "^1.2.2.post1"
setuptools = "^80.9.0"
django-extended-ol = "^1.2.1"
django-allauth = "^65.13.1"
pyjwt = "^2.10.1"
cryptography = "^46.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
